for now, force off Rocksanity when Bucket List or Lost Worlds used
In future, ideally Rocksanity and Bucket List will work well, together.
If desired, Rocksanity could be added to Lost Worlds, as well, but
probably only makes sense in Chronosanity (it only adds 3 of the 5
rock spots, one being in Black Omen, so might not be likely to
be used often).