flake8 config based on existing code, constrain test requirements
Went through code and organized all current flake8 failures and
added to ignore in .flake8. Some are issues that probably should
be addressed soon, others are minor formatting, and some contributors
just might not care about (trying not to be dogmatic). Went with 119
line-length because almost all code in repo has less than that (and
quite a bit has more than 79 characters, which is another common
boundary for python projects).